@@ -12,11 +12,13 @@
     <url>http://topjava.herokuapp.com/</url>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
         <spring.version>5.2.0.RELEASE</spring.version>
+        <spring-data-jpa.version>2.2.0.RELEASE</spring-data-jpa.version>
+        <tomcat.version>9.0.27</tomcat.version>
 
         <!-- Logging -->
         <logback.version>1.2.3</logback.version>
