@@ -7,9 +7,14 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
 
 public class DateTimeUtil {
     private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
+
+    // HSQLDB doesn't support LocalDate.MIN/MAX
+    private static final LocalDate MIN_DATE = LocalDate.of(1, 1, 1);
+    private static final LocalDate MAX_DATE = LocalDate.of(3000, 1, 1);
 
     private DateTimeUtil() {
     }
@@ -26,8 +31,16 @@ public class DateTimeUtil {
         return StringUtils.isEmpty(str) ? null : LocalTime.parse(str);
     }
 
-    public static LocalDateTime createDateTime(@Nullable LocalDate date, LocalDate defaultDate, LocalTime time) {
-        return LocalDateTime.of(date != null ? date : defaultDate, time);
+    public static LocalDateTime getStartInclusive(LocalDate localDate) {
+        return startOfDay(localDate != null ? localDate : MIN_DATE);
+    }
+
+    public static LocalDateTime getEndExclusive(LocalDate localDate) {
+        return startOfDay(localDate != null ? localDate.plus(1, ChronoUnit.DAYS) : MAX_DATE);
+    }
+
+    private static LocalDateTime startOfDay(LocalDate localDate) {
+        return LocalDateTime.of(localDate, LocalTime.MIN);
     }
 }
 
